Move database accessor
